+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Changelog
 0.x.x (?)
 ==================
 
+* Extended DashboardLink to support links to dashboards and urls, as per the docs_
+
+.. _`docs`: https://grafana.com/docs/grafana/latest/dashboards/build-dashboards/manage-dashboard-links/#dashboard-links
+
 * Added ...
 * Added Minimum option for Timeseries
 * Added Maximum option for Timeseries