+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ x.x.x ?
 ==================
 
 * Added `neutral` option for `GaugePanel` (supported by Grafana 9.3.0 - https://github.com/grafana/grafana/discussions/38273)
+* Added support `alias` via the `legendFormat` option for `Target`
 
 0.7.1 2024-01-12
 ==================